Move/split tests as well
