+++ src/BUILD.gn
@@ -40,6 +40,7 @@ if (chip_build_tests) {
       "${chip_root}/src/crypto/tests",
       "${chip_root}/src/inet/tests",
       "${chip_root}/src/system/tests",
+      "${chip_root}/src/transport/raw/tests",
       "${chip_root}/src/transport/retransmit/tests",
       "${chip_root}/src/transport/tests",
     ]
