+++ src/transport/raw/tests/TestTCPDriver.cpp
@@ -22,7 +22,7 @@
  *
  */
 
-#include "TestTransportLayer.h"
+#include "TestRawTransportLayer.h"
 
 #include <nlunit-test.h>
