+++ src/transport/raw/tests/TestUDP.cpp
@@ -21,7 +21,7 @@
  *      This file implements unit tests for the UdpTransport implementation.
  */
 
-#include "TestTransportLayer.h"
+#include "TestRawTransportLayer.h"
 
 #include "NetworkTestHelpers.h"
