+++ src/transport/raw/tests/TestUDPDriver.cpp
@@ -18,17 +18,18 @@
 /**
  *    @file
  *      This file implements a standalone/native program executable
- *      test driver for the CHIP Transport Layer Message header unit
- *      tests.
+ *      test driver for the CHIP core library CHIP Connection tests.
  *
  */
 
-#include "TestTransportLayer.h"
+#include "TestRawTransportLayer.h"
 
 #include <nlunit-test.h>
 
 int main(void)
 {
+    // Generate machine-readable, comma-separated value (CSV) output.
     nlTestSetOutputStyle(OUTPUT_CSV);
-    return TestMessageHeader();
+
+    return (TestUDP());
 }
