+++ src/transport/tests/TestSecureSessionMgr.cpp
@@ -23,11 +23,10 @@
 
 #include "TestTransportLayer.h"
 
-#include "NetworkTestHelpers.h"
-
 #include <core/CHIPCore.h>
 #include <support/CodeUtils.h>
 #include <transport/SecureSessionMgr.h>
+#include <transport/raw/tests/NetworkTestHelpers.h>
 
 #include <nlbyteorder.h>
 #include <nlunit-test.h>
